requirements.txt
transformers
timm
torch
torchvision
scikit-learn
opencv-python
matplotlib
pillow